Filter by forensic STRs [skip ci]
